Implement clinic-veterinary assignment system and finalize CRUD updates
- Moved and refactored AdminClinicController and AdminController
- Added clinic_veterinaries.jsp to manage veterinary-clinic assignments
- Created ClinicVeterinaryRepository and integrated into controller
- Completed CRUD for veterinary and clinic entities
- Fixed veterinary edit page (hidden ID field bug)
- Enabled assignment and removal of veterinarians from clinics
- Updated views, services, and repositories accordingly

+++ src/main/java/com/vetapp/veterinarysystem/controller/UserController.java
@@ -22,7 +22,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public User getUser(@PathVariable int id) {
+    public User getUser(@PathVariable Long id) {
         return userService.getUserById(id);
     }
 
@@ -32,12 +32,12 @@ public class UserController {
     }
 
     @PutMapping("/{id}")
-    public User updateUser(@PathVariable int id, @RequestBody User user) {
+    public User updateUser(@PathVariable Long id, @RequestBody User user) {
         return userService.updateUser(id, user);
     }
 
     @DeleteMapping("/{id}")
-    public void deleteUser(@PathVariable int id) {
+    public void deleteUser(@PathVariable Long id) {
         userService.deleteUser(id);
     }
 }
