+++ src/main/java/com/vetapp/veterinarysystem/repository/ClinicRepository.java
@@ -3,5 +3,5 @@ package com.vetapp.veterinarysystem.repository;
 import com.vetapp.veterinarysystem.model.Clinic;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClinicRepository extends JpaRepository<Clinic, Integer> {
+public interface ClinicRepository extends JpaRepository<Clinic, Long> {
 }
