+++ src/main/java/com/vetapp/veterinarysystem/service/UserService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface UserService {
     List<User> getAllUsers();
-    User getUserById(int id);
+    User getUserById(Long id);
     User createUser(User user);
-    User updateUser(int id, User user);
-    void deleteUser(int id);
+    User updateUser(Long id, User user);
+    void deleteUser(Long id);
 }
